WFE-87 Extract MDSDBlackboard into new dedicated bundle.
