+++ bundles/de.uka.ipd.sdq.workflow.mdsd.blackboard/META-INF/MANIFEST.MF
@@ -1,0 +1,11 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: MDSD Jobs for Palladio Workflow Engine
+Bundle-SymbolicName: de.uka.ipd.sdq.workflow.mdsd.blackboard
+Bundle-Version: 5.1.0.qualifier
+Automatic-Module-Name: de.uka.ipd.sdq.workflow.mdsd.blackboard
+Bundle-RequiredExecutionEnvironment: JavaSE-11
+Bundle-Vendor: sdq.ipd.uka.de
+Export-Package: de.uka.ipd.sdq.workflow.mdsd.blackboard
+Require-Bundle: de.uka.ipd.sdq.workflow;bundle-version="5.1.0";visibility:=reexport,
+ org.eclipse.emf.ecore;bundle-version="2.23.0";visibility:=reexport
