+++ bundles/de.uka.ipd.sdq.workflow.mdsd.blackboard/build.properties
@@ -1,0 +1,4 @@
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .
