+++ bundles/de.uka.ipd.sdq.workflow.mdsd.blackboard/src/de/uka/ipd/sdq/workflow/mdsd/blackboard/ResourceSetPartition.java
