+++ bundles/de.uka.ipd.sdq.workflow.mdsd/META-INF/MANIFEST.MF
@@ -6,7 +6,6 @@ Bundle-Version: 5.1.0.qualifier
 Bundle-Vendor: sdq.ipd.uka.de
 Bundle-RequiredExecutionEnvironment: JavaSE-11
 Export-Package: de.uka.ipd.sdq.workflow.mdsd,
- de.uka.ipd.sdq.workflow.mdsd.blackboard,
  de.uka.ipd.sdq.workflow.mdsd.emf,
  de.uka.ipd.sdq.workflow.mdsd.emf.qvto,
  de.uka.ipd.sdq.workflow.mdsd.oaw,
@@ -39,6 +38,7 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="3.4.0",
  com.google.inject;bundle-version="2.0.0",
  org.eclipse.xtext;bundle-version="2.0.1",
  org.eclipse.xtext.generator;bundle-version="2.0.1",
- de.uka.ipd.sdq.workflow.workbench;bundle-version="1.0.0"
+ de.uka.ipd.sdq.workflow.workbench;bundle-version="1.0.0",
+ de.uka.ipd.sdq.workflow.mdsd.blackboard;bundle-version="5.1.0";visibility:=reexport
 Bundle-Activator: de.uka.ipd.sdq.workflow.mdsd.Activator
 Bundle-ActivationPolicy: lazy
