+++ tests/de.uka.ipd.sdq.workflow.mdsd.blackboard.tests/META-INF/MANIFEST.MF
@@ -1,0 +1,11 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: WFE MDSD Blackboard Tests
+Bundle-SymbolicName: de.uka.ipd.sdq.workflow.mdsd.blackboard.tests
+Bundle-Version: 5.1.0.qualifier
+Fragment-Host: de.uka.ipd.sdq.workflow.mdsd.blackboard;bundle-version="5.1.0"
+Automatic-Module-Name: de.uka.ipd.sdq.workflow.mdsd.blackboard.tests
+Bundle-RequiredExecutionEnvironment: JavaSE-11
+Import-Package: org.junit;version="4.12.0"
+Require-Bundle: org.eclipse.emf.ecore.xmi;bundle-version="2.16.0",
+ org.hamcrest.library;bundle-version="1.3.0"
