+++ tests/de.uka.ipd.sdq.workflow.mdsd.blackboard.tests/build.properties
@@ -1,0 +1,5 @@
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .,\
+               models/
