+++ tests/de.uka.ipd.sdq.workflow.mdsd.blackboard.tests/models/test.ecore
@@ -1,0 +1,3 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<ecore:EPackage xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:ecore="http://www.eclipse.org/emf/2002/Ecore"
+    name="test"/>
